+++ src/microControl/Dispenser.h
@@ -8,9 +8,9 @@ right, left, center
 
 class Dispenser {
 private:
-	int amountAvailable;
-	int outputCenter, outputLeft, outputRight;
-	int pin = 6;
+	uint8_t amountAvailable;
+	uint8_t outputCenter, outputLeft, outputRight;
+	uint8_t pin = 6;
 	Servo servoDispenser;
  	Dispenser();
     Dispenser &operator=(const Dispenser &);
@@ -19,10 +19,10 @@ public:
 		static Dispenser singletonInstance;
 		return singletonInstance;
 	};
-	void setAmountAvailable(int iAmount);
-	int getAmountAvailable();
-	int updateAmountAvailable(int drop);
-	void dispenseDirection(DispenserDirection direction, int amount = 1);
+	void setAmountAvailable(uint8_t iAmount);
+	uint8_t getAmountAvailable();
+	uint8_t updateAmountAvailable(uint8_t drop);
+	void dispenseDirection(DispenserDirection direction, uint8_t amount = 1);
 	bool kitsAvailable();
 }; 
 #endif
